squashme: move the cache sync to the watcher
