+++ test/bases/renku_data_services/data_api/test_notebooks.py
@@ -259,6 +259,7 @@ class TestNotebooks(ClusterRequired):
             handler=k8s_object_handler(app_config.nb_config.k8s_db_cache, app_config.metrics, app_config.rp_repo),
             clusters={c.id: c for c in clusters},
             kinds=[JUPYTER_SESSION_GVK],
+            db_cache=app_config.nb_config.k8s_db_cache,
         )
         asyncio.create_task(watcher.start())
         yield
